+++ test/sea_test.exs
@@ -10,12 +10,9 @@ defmodule SeaTest do
 
       defstruct [:number_as_string]
 
-      emit_to(SeaTest.Mod1)
-      emit_to(SeaTest.{Mod2, Mod3})
-      emit_to([SeaTest.Mod4, SeaTest.Mod5])
-      emit_within(SeaTest.Mod6)
-      emit_within(SeaTest.{Mod7, Mod8})
-      emit_within([SeaTest.Mod9, SeaTest.ModA])
+      emit_to SeaTest.Mod1
+      emit_to SeaTest.{Mod2, Mod3}
+      emit_to [SeaTest.Mod4, SeaTest.Mod5]
 
       def build(1) do
         %__MODULE__{number_as_string: "one"}
@@ -67,51 +64,6 @@ defmodule SeaTest do
       end
     end
 
-    defmodule Mod6.SampleObserver do
-      use Sea.Observer
-
-      @impl true
-      def handle_signal(%SampleSignal{number_as_string: str}) do
-        IO.puts("Mod6 got #{str}")
-      end
-    end
-
-    defmodule Mod7.SampleObserver do
-      use Sea.Observer
-
-      @impl true
-      def handle_signal(%SampleSignal{number_as_string: str}) do
-        IO.puts("Mod7 got #{str}")
-      end
-    end
-
-    defmodule Mod8.SampleObserver do
-      use Sea.Observer
-
-      @impl true
-      def handle_signal(%SampleSignal{number_as_string: str}) do
-        IO.puts("Mod8 got #{str}")
-      end
-    end
-
-    defmodule Mod9.SampleObserver do
-      use Sea.Observer
-
-      @impl true
-      def handle_signal(%SampleSignal{number_as_string: str}) do
-        IO.puts("Mod9 got #{str}")
-      end
-    end
-
-    defmodule ModA.SampleObserver do
-      use Sea.Observer
-
-      @impl true
-      def handle_signal(%SampleSignal{number_as_string: str}) do
-        IO.puts("ModA got #{str}")
-      end
-    end
-
     assert capture_io(fn ->
              SampleSignal.emit(1)
            end) == """
@@ -120,11 +72,6 @@ defmodule SeaTest do
            Mod3 got one
            Mod4 got one
            Mod5 got one
-           Mod6 got one
-           Mod7 got one
-           Mod8 got one
-           Mod9 got one
-           ModA got one
            """
   end
 
@@ -134,7 +81,7 @@ defmodule SeaTest do
 
       defstruct []
 
-      emit_to(SeaTest.MockableObserver)
+      emit_to SeaTest.MockableObserver
 
       def build(:empty), do: %__MODULE__{}
     end